# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * A professional REST API for user management, posts, and comments. Built with Node.js, Express, and PostgreSQL. Supports full CRUD operations, pagination, filtering, search, and structured error handling.

 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  bio?: string | null;
  avatarUrl?: string | null;
  isActive: boolean;
  postCount: number;
  createdAt: Date;
  updatedAt: Date;
}